requirements.txt
Django
djangorestframework
black
flake8
pytest
pytest-django
